Refactor MyGames and MyProjects card design, move into same css file
