+++ client/src/pages/MyGames&MyProjects/MyProjects.jsx
@@ -9,7 +9,7 @@ import {
   Chip,
   Stack,
 } from "@mui/material";
-import styles from "./MyProjects.module.css";
+import styles from "./CardDesign.module.css";
 import { projects } from "../data/projectData.js";
 
 function MyProjects() {
