--- client/src/pages/MyGames/MyGames.module.css
@@ -1,40 +1,0 @@
-.wrapper {
-  margin-top: 4rem;
-  margin-bottom: 8rem;
-}
-
-.title {
-  text-align: center;
-  font-weight: bold;
-  color: var(--color-secondary);
-}
-
-.card {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  transition: transform 0.3s ease, box-shadow 0.3s ease;
-  background-color: white;
-  cursor: default;
-}
-
-.card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-}
-
-.cardTitle {
-  color: var(--color-primary);
-  font-weight: 600;
-  margin-bottom: 0.5rem;
-}
-
-.cardText {
-  color: var(--color-text-primary);
-}
-
-.button {
-  margin-left: 0.5rem;
-}
